rename hash.h to zobrist.h
